Use .ratchet.toml rather than pyproject.toml for configuration
The pyproject.toml file should not be written to by tools, as it is
intended for project authors.

Writing back to a configuration file would potentially re-format it,
mess with diffs, and undo any specific changes or comments made by the
authors.

Instead, we provide our own config file that is owned and managed by
this tool.